+++ schematics/core/files/core/index.ts
@@ -1,0 +1,19 @@
+// Domain Layer - Base Classes
+export { BaseEntity } from './domain/base/base.entity';
+export { AggregateRoot } from './domain/base/aggregate-root';
+export { ValueObject } from './domain/base/value-object';
+
+// Domain Layer - Repositories
+export { BaseRepository } from './domain/repositories/base-repository.interface';
+
+// Domain Layer - Events
+export { DomainEvent } from './domain/events/domain-event';
+export { DomainEventHandler } from './domain/events/domain-event-handler.interface';
+
+// Application Layer
+export { BaseUseCase } from './application/use-cases/base-use-case.interface';
+export { BaseMapper } from './application/mappers/base-mapper.interface';
+
+// Common - Result/Either Monads
+export { Result } from './common/result/result';
+export { Either } from './common/result/either';